+++ NEWS.md
@@ -2,6 +2,7 @@
 
 - Add `use_detect_data` and `plot_use_detect` functions for habitat use.
 - Add `combine_sections_lex_data` function for simpler interface for combining sections.
+- Add `last_section_data` to get number of fish last detected at each section.
 
 ## v0.0.8